- Updated to version 1.13.2

+++ RockWeb/Blocks/WorkFlow/WorkflowTypeDetail.ascx.cs
@@ -323,15 +323,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 LoadStateDetails( workflowType, rockContext );
 
                 // clone the workflow type
-                var newWorkflowType = workflowType.Clone( false );
-                newWorkflowType.CreatedByPersonAlias = null;
-                newWorkflowType.CreatedByPersonAliasId = null;
-                newWorkflowType.CreatedDateTime = RockDateTime.Now;
-                newWorkflowType.ModifiedByPersonAlias = null;
-                newWorkflowType.ModifiedByPersonAliasId = null;
-                newWorkflowType.ModifiedDateTime = RockDateTime.Now;
-                newWorkflowType.Id = 0;
-                newWorkflowType.Guid = Guid.NewGuid();
+                var newWorkflowType = workflowType.CloneWithoutIdentity();
                 newWorkflowType.IsSystem = false;
                 newWorkflowType.Name = workflowType.Name + " - Copy";
 
@@ -467,8 +459,27 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                                     newFormAttribute.Attribute = new Rock.Model.Attribute
                                     {
                                         Guid = guidXref[formAttribute.Attribute.Guid],
-                                        Name = formAttribute.Attribute.Name
+                                        Name = formAttribute.Attribute.Name,
+                                        FieldTypeId = formAttribute.Attribute.FieldTypeId,
+                                        AttributeQualifiers = formAttribute.Attribute.AttributeQualifiers
                                     };
+
+                                    if ( newFormAttribute.FieldVisibilityRules != null )
+                                    {
+                                        var visibilityRules = newFormAttribute.FieldVisibilityRules.Clone();
+
+                                        foreach ( var rule in visibilityRules.RuleList )
+                                        {
+                                            if ( rule.ComparedToFormFieldGuid != null && guidXref.ContainsKey( rule.ComparedToFormFieldGuid.Value ) )
+                                            {
+                                                rule.ComparedToFormFieldGuid = guidXref[rule.ComparedToFormFieldGuid.Value];
+                                            }
+                                        }
+
+                                        newFormAttribute.FieldVisibilityRules = visibilityRules;
+                                    }
+
+
                                     newWorkflowForm.FormAttributes.Add( newFormAttribute );
                                 }
                             }
@@ -607,6 +618,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 workflowType.ProcessingIntervalSeconds = null;
             }
 
+            workflowType.MaxWorkflowAgeDays = tbMaximumWorkflowAge.Text.AsIntegerOrNull();
             workflowType.LogRetentionPeriod = tbLogRetention.Text.AsIntegerOrNull();
             workflowType.CompletedWorkflowRetentionPeriod = tbCompletedRetention.Text.AsIntegerOrNull();
             workflowType.IsPersisted = cbIsPersisted.Checked;
@@ -726,6 +738,8 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
             }
             rockContext.SaveChanges();
 
+            var formBuilderEntityTypeId = EntityTypeCache.GetId<Rock.Workflow.Action.FormBuilder>();
+
             // add or update WorkflowActivityTypes(and Actions) that are assigned in the UI
             int workflowActivityTypeOrder = 0;
             foreach ( var editorWorkflowActivityType in ActivityTypesState )
@@ -753,12 +767,6 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     SaveAttributes( new WorkflowActivity().TypeId, "ActivityTypeId", workflowActivityType.Id.ToString(), ActivityAttributesState[workflowActivityType.Guid], rockContext );
                 }
 
-                // Because the SaveAttributes above may have flushed the cached entity attribute cache, and it would get loaded again with
-                // a different context, manually reload the cache now with our context to prevent a database lock conflict (when database is 
-                // configured without snapshot isolation turned on)
-                EntityAttributesCache.Remove();
-                EntityAttributesCache.Get( rockContext );
-
                 int workflowActionTypeOrder = 0;
                 foreach ( var editorWorkflowActionType in editorWorkflowActivityType.ActionTypes )
                 {
@@ -781,14 +789,16 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     workflowActionType.AttributeValues = editorWorkflowActionType.AttributeValues;
                     workflowActionType.Order = workflowActionTypeOrder++;
 
-                    if ( workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
+                    var isFormBuilderAction = workflowActionType.EntityTypeId == formBuilderEntityTypeId;
+
+                    if ( !isFormBuilderAction && workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
                     {
                         // Form removed
                         workflowFormService.Delete( workflowActionType.WorkflowForm );
                         workflowActionType.WorkflowForm = null;
                     }
 
-                    if ( editorWorkflowActionType.WorkflowForm != null )
+                    if ( !isFormBuilderAction && editorWorkflowActionType.WorkflowForm != null )
                     {
                         if ( workflowActionType.WorkflowForm == null )
                         {
@@ -831,6 +841,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                             formAttribute.HideLabel = editorAttribute.HideLabel;
                             formAttribute.PreHtml = editorAttribute.PreHtml;
                             formAttribute.PostHtml = editorAttribute.PostHtml;
+                            formAttribute.FieldVisibilityRules = editorAttribute.FieldVisibilityRules;
                         }
                     }
                 }
@@ -909,7 +920,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
         {
             ParseControls();
 
-            Guid attributeGuid = (Guid)e.RowKeyValue;
+            Guid attributeGuid = ( Guid ) e.RowKeyValue;
             ShowAttributeEdit( attributeGuid );
         }
 
@@ -938,7 +949,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
         {
             ParseControls();
 
-            Guid attributeGuid = (Guid)e.RowKeyValue;
+            Guid attributeGuid = ( Guid ) e.RowKeyValue;
             AttributesState.RemoveEntity( attributeGuid );
 
             BindAttributesGrid();
@@ -1294,7 +1305,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
 
             if ( workflowType.IsSystem )
             {
-                readOnly = true; 
+                readOnly = true;
                 nbEditModeMessage.Heading = "Information";
                 nbEditModeMessage.Text = EditModeMessage.ReadOnlySystem( WorkflowType.FriendlyTypeName );
             }
@@ -1436,10 +1447,11 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 tbProcessingInterval.Text = string.Empty;
             }
 
+            tbMaximumWorkflowAge.Text = workflowType.MaxWorkflowAgeDays.ToStringSafe();
             tbLogRetention.Text = workflowType.LogRetentionPeriod.ToStringSafe();
             tbCompletedRetention.Text = workflowType.CompletedWorkflowRetentionPeriod.ToStringSafe();
             cbIsPersisted.Checked = workflowType.IsPersisted;
-            ddlLoggingLevel.SetValue( (int)workflowType.LoggingLevel );
+            ddlLoggingLevel.SetValue( ( int ) workflowType.LoggingLevel );
             tbIconCssClass.Text = workflowType.IconCssClass;
             ceSummaryViewText.Text = workflowType.SummaryViewText;
             ceNoActionMessage.Text = workflowType.NoActionMessage;
@@ -1703,10 +1715,11 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 // Add any new attributes
                 foreach ( var attribute in attributes )
                 {
-                    if ( !formAttributes.Select( a => a.Attribute.Guid ).Contains( attribute.Key ) )
+                    var formAttribute = formAttributes.Where( fa => fa.Attribute.Guid == attribute.Key ).FirstOrDefault();
+                    if ( formAttribute == null )
                     {
-                        var formAttribute = new WorkflowActionFormAttribute();
-                        formAttribute.Attribute = new Rock.Model.Attribute { Guid = attribute.Key, Name = attribute.Value.Name };
+                        formAttribute = new WorkflowActionFormAttribute();
+                        formAttribute.Attribute = new Attribute { Guid = attribute.Key, Name = attribute.Value.Name, FieldTypeId = attribute.Value.FieldTypeId };
                         formAttribute.Guid = Guid.NewGuid();
                         formAttribute.Order = formAttributes.Any() ? formAttributes.Max( a => a.Order ) + 1 : 0;
                         formAttribute.IsVisible = false;
@@ -1717,6 +1730,9 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                         formAttribute.PostHtml = string.Empty;
                         formAttributes.Add( formAttribute );
                     }
+
+                    // make sure formAttribute AttributeQualifiers matches the current qualifiers if they have been added/edited
+                    formAttribute.Attribute.AttributeQualifiers = attribute.Value.AttributeQualifiers.ToList();
                 }
             }
             control.SetWorkflowActionType( actionType, attributes );
